requirements.txt
lumibot
alpaca-py>=0.42.0
pandas>=2.2.0
numpy>=1.20.0
yfinance>=0.2.61
python-dotenv
requests
agno
openai
anthropic
mcp
schedule
pytest
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
psycopg[binary]>=3.1
